+++ gltf-loader/Cargo.toml
@@ -10,3 +10,4 @@ serde_json = "1.0"
 serde_derive = "1.0"
 serde_repr = "0.1"
 bedrock = { git = "https://github.com/Pctg-x8/bedrock" }
+peridot = { path = ".." }
